+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
-## 0.0.2
+## 0.0.3
+
+### Features
+
+- Add `with_executor` in ReplayOptions.
+
+## 0.0.2 - 2025-12-26
 
 ### Features